+++ src/Sulu/Bundle/MediaBundle/Tests/Functional/Controller/MediaControllerTest.php
@@ -201,11 +201,13 @@ class MediaControllerTest extends SuluTestCase
         $this->audioType->setName('audio');
         $this->audioType->setDescription('This is a video');
 
+        $tagRepository = $this->getContainer()->get('sulu.repository.tag');
+
         // create some tags
-        $tag1 = new Tag();
+        $tag1 = $tagRepository->createNew();
         $tag1->setName('Tag 1');
 
-        $tag2 = new Tag();
+        $tag2 = $tagRepository->createNew();
         $tag2->setName('Tag 2');
 
         $this->em->persist($tag1);