+++ src/Sulu/Bundle/MediaBundle/Tests/Functional/Controller/MediaRedirectControllerTest.php
@@ -190,11 +190,13 @@ class MediaRedirectControllerTest extends SuluTestCase
         $this->videoType->setName('video');
         $this->videoType->setDescription('This is a video');
 
+        $tagRepository = $this->getContainer()->get('sulu.repository.tag');
+
         // create some tags
-        $tag1 = new Tag();
+        $tag1 = $tagRepository->createNew();
         $tag1->setName('Tag 1');
 
-        $tag2 = new Tag();
+        $tag2 = $tagRepository->createNew();
         $tag2->setName('Tag 2');
 
         $this->em->persist($tag1);